requirements.txt
mysql-connector-python
flask
boto3
botocore
